+++ lib/plausible/clickhouse_event_v2.ex
@@ -70,7 +70,8 @@ defmodule Plausible.ClickhouseEventV2 do
         :revenue_source_amount,
         :revenue_source_currency,
         :revenue_reporting_amount,
-        :revenue_reporting_currency
+        :revenue_reporting_currency,
+        :interactive?
       ]
     )
     |> validate_required([:name, :site_id, :hostname, :pathname, :user_id, :timestamp])