+++ test/plausible_web/controllers/api/external_stats_controller/query_imported_test.exs
@@ -683,25 +683,25 @@ defmodule PlausibleWeb.Api.ExternalStatsController.QueryImportedTest do
                breakdown_and_first.("visit:source")
     end
 
-    for goal_name <- Plausible.Goals.SystemGoals.goals_with_url() do
-      test "returns url breakdown for #{goal_name} goal", %{conn: conn, site: site} do
-        insert(:goal, event_name: unquote(goal_name), site: site)
+    for event_name <- Plausible.Event.SystemEvents.events_with_url_prop() do
+      test "returns url breakdown for #{event_name} goal", %{conn: conn, site: site} do
+        insert(:goal, event_name: unquote(event_name), site: site)
         site_import = insert(:site_import, site: site)
 
         populate_stats(site, site_import.id, [
           build(:event,
-            name: unquote(goal_name),
+            name: unquote(event_name),
             "meta.key": ["url"],
             "meta.value": ["https://one.com"]
           ),
           build(:imported_custom_events,
-            name: unquote(goal_name),
+            name: unquote(event_name),
             visitors: 2,
             events: 5,
             link_url: "https://one.com"
           ),
           build(:imported_custom_events,
-            name: unquote(goal_name),
+            name: unquote(event_name),
             visitors: 5,
             events: 10,
             link_url: "https://two.com"
@@ -721,7 +721,7 @@ defmodule PlausibleWeb.Api.ExternalStatsController.QueryImportedTest do
             "date_range" => "all",
             "dimensions" => ["event:props:url"],
             "filters" => [
-              ["is", "event:goal", [unquote(goal_name)]]
+              ["is", "event:goal", [unquote(event_name)]]
             ],
             "include" => %{"imports" => true}
           })
@@ -736,25 +736,25 @@ defmodule PlausibleWeb.Api.ExternalStatsController.QueryImportedTest do
       end
     end
 
-    for goal_name <- Plausible.Goals.SystemGoals.goals_with_path() do
-      test "returns path breakdown for #{goal_name} goal", %{conn: conn, site: site} do
-        insert(:goal, event_name: unquote(goal_name), site: site)
+    for event_name <- Plausible.Event.SystemEvents.events_with_path_prop() do
+      test "returns path breakdown for #{event_name} goal", %{conn: conn, site: site} do
+        insert(:goal, event_name: unquote(event_name), site: site)
         site_import = insert(:site_import, site: site)
 
         populate_stats(site, site_import.id, [
           build(:event,
-            name: unquote(goal_name),
+            name: unquote(event_name),
             "meta.key": ["path"],
             "meta.value": ["/one"]
           ),
           build(:imported_custom_events,
-            name: unquote(goal_name),
+            name: unquote(event_name),
             visitors: 2,
             events: 5,
             path: "/one"
           ),
           build(:imported_custom_events,
-            name: unquote(goal_name),
+            name: unquote(event_name),
             visitors: 5,
             events: 10,
             path: "/two"
@@ -774,7 +774,7 @@ defmodule PlausibleWeb.Api.ExternalStatsController.QueryImportedTest do
             "date_range" => "all",
             "dimensions" => ["event:props:path"],
             "filters" => [
-              ["is", "event:goal", [unquote(goal_name)]]
+              ["is", "event:goal", [unquote(event_name)]]
             ],
             "include" => %{"imports" => true}
           })